+++ src/plugins/lv2/lv2ui.cpp
@@ -53,8 +53,17 @@ struct ParameterListener final : public Parameter::Observer {
 		writeFunc(parameter.getId(), parameter.getValue());
 	}
 
+	void parameterBeginEdit(const Parameter &parameter) final {
+		touchFunc(parameter.getId(), true);
+	}
+
+	void parameterEndEdit(const Parameter &parameter) final {
+		touchFunc(parameter.getId(), false);
+	}
+
 	PresetController *presetController;
 	std::function<void(int, float)> writeFunc;
+	std::function<void(int, bool)> touchFunc;
 	bool active{false};
 };
 
@@ -69,6 +78,7 @@ struct lv2_ui {
 
 	LV2_Atom_Forge forge;
 	LV2_URID_Map *map;
+	LV2UI_Touch *touch {nullptr};
 	LV2UI_Write_Function _write_function;
 	LV2UI_Controller _controller;
 
@@ -171,6 +181,8 @@ lv2_ui_instantiate(const LV2UI_Descriptor*         /*descriptor*/,
 			ui->parent = reinterpret_cast<LV2UI_Widget>((*f)->data);
 		if (!strcmp((*f)->URI, LV2_URID__map))
 			ui->map = reinterpret_cast<LV2_URID_Map *>((*f)->data);
+		if (!strcmp((*f)->URI, LV2_UI__touch))
+			ui->touch = reinterpret_cast<LV2UI_Touch *>((*f)->data);
 		if (!strcmp((*f)->URI, LV2_UI__resize))
 			resize = reinterpret_cast<LV2UI_Resize *>((*f)->data);
 	}
@@ -203,6 +215,11 @@ lv2_ui_instantiate(const LV2UI_Descriptor*         /*descriptor*/,
 	ui->parameterListener = std::make_unique<ParameterListener>(&ui->presetController, [=] (int idx, float value) {
 		write_function(controller, PORT_FIRST_PARAMETER + idx, sizeof(float), 0, &value);
 	});
+
+	ui->parameterListener->touchFunc = [=](int idx, bool grabbed) {
+		if (!ui->touch) return;
+		(*ui->touch->touch)(ui->touch->handle, PORT_FIRST_PARAMETER + idx, grabbed);
+	};
 
 	ui->mainComponent = std::make_unique<MainComponent>(&ui->presetController);
 	ui->mainComponent->sendProperty = [ui] (const char *name, const char *value) {lv2helper(ui).send(name, value);};